Support Lazy-loading `<turbo-frame>` contents
Add support for declaring `[loading="lazy"]` (in the style of
[`img[loading="lazy"]`][img-loading]). When both `src` and
`loading="lazy"` are declared on the frame, navigation will be deferred
until the element is visible in the document.

As an example, consider a `<details>`/`<summary>` pairing:

```html
<details>
  <summary>Lazy load</summary>

  <turbo-frame id="frame" src="/lazy-loaded.html" loading="lazy"></turbo-frame>
</details>
```

On initial load, the frame's contents will be empty. When the
`<summary>` element is clicked and the contents are visible, the
`<turbo-frame>` element will navigate to the URL specified in the `src`
element.

By default, `<turbo-frame>` elements that omit the `loading` attribute
will be treated as if they've declared `loading="eager"`.

[img-loading]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Img#attr-loading